dev environmnet set --- entities done --- sql script done (to review)

+++ backend/src/main/java/com/proposito365/app/models/User.java
@@ -13,7 +13,7 @@ public class User {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="id")
-	private int id;
+	private Long id;
 
 	@Column(name="email")
 	private String email;
@@ -22,17 +22,17 @@ public class User {
 
 	public User() {}
 
-	public User(int id, String email, String password) {
+	public User(Long id, String email, String password) {
 		this.id = id;
 		this.email = email;
 		this.password = password;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
